+++ src/rules/property-blacklist/index.js
@@ -15,7 +15,7 @@ export const messages = ruleMessages(ruleName, {
   rejected: (property) => `Unexpected property "${property}"`,
 })
 
-export default function (blacklist) {
+function rule(blacklist) {
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
       actual: blacklist,
@@ -38,3 +38,7 @@ export default function (blacklist) {
     })
   }
 }
+
+rule.primaryOptionArray = true
+
+export default rule